feat: incident retro pure helpers out of google workspace
